pacify clang
git-svn-id: https://svn.r-project.org/R/trunk@59990 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/grDevices/src/cairo/cairoBM.c
@@ -93,11 +93,11 @@ static Rboolean BM_Open(pDevDesc dd, pX11Desc xd, int width, int height)
             return FALSE;
         }
 #endif
-        xd->cs = cairo_image_surface_create(CAIRO_FORMAT_ARGB32, (double)xd->windowWidth, (double)xd->windowHeight);
+        xd->cs = cairo_image_surface_create(CAIRO_FORMAT_ARGB32, xd->windowWidth, xd->windowHeight);
     }
     else if (xd->type == PNGdirect)
     {
-        xd->cs = cairo_image_surface_create(CAIRO_FORMAT_ARGB32, (double)xd->windowWidth, (double)xd->windowHeight);
+        xd->cs = cairo_image_surface_create(CAIRO_FORMAT_ARGB32, xd->windowWidth, xd->windowHeight);
     }
     else if (xd->type == SVG || xd->type == PDF || xd->type == PS)
     {
